Added settings processing for client and ppm.

+++ lib/settings.rb
@@ -8,7 +8,13 @@ def configuration_file_name
 end
 
 def read_configuration
-  eval File.open(CONFIG_FILE_NAME).read
+  if File.exists? CONFIG_FILE_NAME
+    eval File.open(CONFIG_FILE_NAME).read
+    if not conf.ppm.base_path.nil? and File.exists? "#{conf.ppm.base_path}/#{CONFIG_FILE_NAME}"
+      eval File.open("#{conf.ppm.base_path}/#{CONFIG_FILE_NAME}").read
+      eval File.open(CONFIG_FILE_NAME).read
+    end
+  end
 end
 
 def conf
